backend/pyproject.toml
[tool.poetry]
name = "reparatur-backend"
version = "0.1.0"
description = "Backend for reparatur app"
authors = ["Hamza Abdelazez <hamzasobhi98@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "api", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.41"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
shapely = "^2.1.1"
pydantic = "^2.11.1"
python-dotenv = "^1.0.1"
poetry-core = "^2.1.3"
pydantic-settings = "^2.9.1"
dataclasses-json = "^0.6.7"
passlib = {version = ">=1.7.4", extras = ["bcrypt"]}
python-multipart = "^0.0.20"
pgvector = "^0.4.1"
bcrypt = ">=4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
